creat example network configuration application

+++ scripts/run_app.sh
@@ -8,10 +8,12 @@ if [ -z $FID_FOLDER ]; then
 	PID_FOLDER=/tmp/
 fi
 
-if [ $1 = "start" ] ; then
-	start-stop-daemon --start --oknodo --make-pidfile --pidfile $PID_FOLDER/app_$INSNAME.pid --chdir /tmp/apps/$INSNAME --background --startas /usr/bin/lua -- start.lua $INSNAME $4 $5 $6 $7
-else
-	start-stop-daemon --stop --oknodo --pidfile $PID_FOLDER/app_$INSNAME.pid --retry 5
-	rm $PID_FOLDER/app_$INSNAME.pid
+# Only run the application if main.lua exists
+if [-f /tmp/apps/main.lua ]; then
+	if [ $1 = "start" ] ; then
+		start-stop-daemon --start --oknodo --make-pidfile --pidfile $PID_FOLDER/app_$INSNAME.pid --chdir /tmp/apps/$INSNAME --background --startas /usr/bin/lua -- start.lua $INSNAME $4 $5 $6 $7
+	else
+		start-stop-daemon --stop --oknodo --pidfile $PID_FOLDER/app_$INSNAME.pid --retry 5
+		rm $PID_FOLDER/app_$INSNAME.pid
+	fi
 fi
-
